requirements.txt
# MCP Server Dependencies
mcp>=1.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
python-dotenv>=1.0.0

# Development Dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
